simplified logging and object packaging
